# web/src/lib/api/generated/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Athenaeum HTTP API",
    "description": "Session cookie (browser): sign in via POST /api/auth/login; cookies are sent automatically.\n\nHTTP Basic Auth: Authorization: Basic base64(username:password) — useful for OPDS clients and scripts.\n\nAPI key: Authorization: Bearer ath_\u003ctoken\u003e or X-API-Key: ath_\u003ctoken\u003e — create keys in Settings -\u003e API.\n\nJSON endpoints accept and return application/json unless noted.\n\nFile endpoints stream binary content with appropriate Content-Type and Accept-Ranges.\n\nMutating browser requests require the X-CSRF-Token header (not needed for Basic Auth or API keys).",
    "version": "dev"
  },
  "servers": [
    {
      "url": "/",
      "description": "This server"
    }
  ],
  "paths": {
...
    "/api/admin/server": {
      "get": {
        "tags": [
          "Admin"
        ],
        "summary": "Server settings (metrics, proxy, CSP, CORS)",
        "description": "Auth: admin",
        "operationId": "GET__api_admin_server",
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "sessionCookie": []
          },
          {
            "basicAuth": []
          },
          {
            "apiKeyHeader": []
          },
          {
            "bearerAuth": []
          }
        ]
      },
      "put": {
        "tags": [
          "Admin"
        ],
        "summary": "Update server settings",
        "description": "Auth: admin",
        "operationId": "PUT__api_admin_server",
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "sessionCookie": []
          },
          {
            "basicAuth": []
          },
          {
            "apiKeyHeader": []
          },
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/books/{id}": {
      "get": {
        "tags": [
          "Books"
        ],
        "summary": "Book detail",
        "description": "Auth: required",
        "operationId": "GET__api_books__id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "sessionCookie": []
          },
          {
            "basicAuth": []
          },
          {
            "apiKeyHeader": []
          },
          {
            "bearerAuth": []
          }
        ]
      },
      "put": {
        "tags": [
          "Books"
        ],
        "summary": "Edit metadata",
        "description": "Auth: admin",
        "operationId": "PUT__api_books__id",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "sessionCookie": []
          },
          {
            "basicAuth": []
          },
          {
            "apiKeyHeader": []
          },
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/books/{id}/cover": {
      "get": {
        "tags": [
          "Books"
        ],
        "summary": "Cover image",
        "description": "Auth: required",
        "operationId": "GET__api_books__id__cover",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "sessionCookie": []
          },
          {
            "basicAuth": []
          },
          {
            "apiKeyHeader": []
          },
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/books/{id}/favorite": {
      "put": {
        "tags": [
          "Books"
        ],
        "summary": "Toggle favorite",
        "description": "Auth: required\nBody: {\"favorite\":true}",
        "operationId": "PUT__api_books__id__favorite",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "sessionCookie": []
          },
          {
            "basicAuth": []
          },
          {
            "apiKeyHeader": []
          },
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/favorites": {
      "get": {
        "tags": [
          "Books"
        ],
        "summary": "Favorite book IDs",
        "description": "Auth: required",
        "operationId": "GET__api_favorites",
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "security": [
          {
            "sessionCookie": []
          },
          {
            "basicAuth": []
          },
          {
            "apiKeyHeader": []
          },
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
    "/api/i18n/locales": {
      "get": {
        "tags": [
          "Health and auth"
        ],
        "summary": "List available UI locales",
        "operationId": "GET__api_i18n_locales",
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/i18n/template": {
      "get": {
        "tags": [
          "Health and auth"
        ],
        "summary": "Translation template (keys with empty values)",
        "operationId": "GET__api_i18n_template",
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/i18n/{locale}": {
      "get": {
        "tags": [
          "Health and auth"
        ],
        "summary": "Messages for one locale",
        "operationId": "GET__api_i18n__locale",
        "parameters": [
          {
            "name": "locale",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "apiKeyHeader": {
        "type": "apiKey",
        "name": "X-API-Key",
        "in": "header",
        "description": "API key (ath_…)"
      },
      "basicAuth": {
        "type": "http",
        "scheme": "basic",
        "description": "Username and password (OPDS and scripts)"
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "API key as Bearer token"
      },
      "sessionCookie": {
        "type": "apiKey",
        "name": "access_token",
        "in": "cookie",
        "description": "Browser session after POST /api/auth/login"
      }
    }
  },
...
}